Refactor database_latest_package and co
